+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 with respect to its command line interface and HTTP interface.
 
-## [Unreleased](//github.com/opentable/sous/compare/0.5.85...HEAD)
+## [0.5.86](//github.com/opentable/sous/compare/0.5.85...0.5.86)
 ### Added
 * Client: `sous manifest edit` covers 90% of the manifest get/set use case, more easily.
 * Server: more flexible, agile logging API.